feat: adicionar sistema de carrinho por usuário com autenticação JWT - Criar entidade CartItem com relação User e Food - Criar CartRepository com métodos findByUser e deleteByUser - Criar CartController com endpoints CRUD (/cart) - Criar DTOs CartItemRequestDTO e CartItemResponseDTO - Adicionar validação de propriedade Long (equals -> ==) - Criar migration V4 para adicionar user_id na tabela cart_items

+++ src/main/java/com/example/cardapio/food/FoodRequestDTO.java
@@ -11,5 +11,5 @@ public record FoodRequestDTO(
     String image,
     @NotNull(message = "Preço é obrigatório")
     @Positive(message = "Preço deve ser positivo")
-    Integer price
+    Double price
 ) {}
