+++ src/main/java/com/example/cardapio/food/FoodResponseDTO.java
@@ -1,6 +1,6 @@
 package com.example.cardapio.food;
 
-public record FoodResponseDTO(long id,String title,String image,Integer price) {
+public record FoodResponseDTO(long id,String title,String image,Double price) {
     public FoodResponseDTO(Food food){
         this(food.getId(), food.getTitle(), food.getImage(), food.getPrice());
     }
